refactor: split skill_creator.py God Object into focused modules
693-line CoworkSkillCreator split into three files:

- generator/skill_project_scanner.py  (ProjectContextScanner)
  Project filesystem scanning: tech stack detection, signal detection,
  project structure analysis, README sufficiency check.

- generator/skill_content_renderer.py  (SkillContentRenderer)
  Skill markdown generation: AI route, Jinja2 template, inline fallback,
  trigger/signal formatting helpers.

- generator/skill_creator.py  (CoworkSkillCreator, now ~250 LOC)
  Thin orchestrator: wires scanner + renderer + quality + metadata builder.
  Keeps backward-compat delegates for _validate_quality, _generate_content,
  _generate_with_jinja2, and detect_skill_needs (patched by existing tests).

593 tests pass, 11 skipped. Zero regressions.
